requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
torch>=2.0.0
yfinance>=0.2.0
alpaca-trade-api>=3.0.0
oandapyV20>=0.6.0
python-dotenv>=1.0.0
markdown>=3.4.0
weasyprint>=59.0
pyyaml>=6.0

